fix routines but break exit because threading
